Moved SurfaceState into own file
