+++ src/types/surface/mod.rs
@@ -1,3 +1,5 @@
 pub mod surface;
+pub mod surface_state;
 
 pub use self::surface::*;
+pub use self::surface_state::*;
